==> notebooks/0.1_data_preprocessing_FullName.ipynb <==
### Full Name Dataset (Train/Validation/Test)

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split

df = pd.read_csv('data/fl_reg_name_race_2022.csv.gz')
df.shape

# Remove NA first/last
df.dropna(subset=['name_first', 'name_last'], inplace=True)
print("Size after dropping missing first and last names:", df.shape)

# We assume unknown as missing at random
sdf = df[df.race.isin(['unknown']) == False]
print("Size after dropping unknown:", sdf.shape)
del df

# Drop cases where last name is less than 2 chars
sdf = sdf.drop(sdf[sdf['name_last'].str.len() < 2].index)
print("Size after dropping last names less than 2 chars:", sdf.shape)

# Full Name
sdf['name_first'] = sdf.name_first.str.strip().str.title()
sdf['name_last'] = sdf.name_last.str.strip().str.title()
sdf['full_name'] = sdf['name_last'] + ' ' + sdf['name_first']
# Remove special chars
sdf['full_name'] = sdf['full_name'].str.replace("[^a-zA-Z' -]", '', regex=True)

# recode race
mapping = {'multi_racial': 'other', 'native_indian': 'other'}
sdf['race'] = sdf['race'].replace(mapping)

# Summing the count of each name & race combination
gdf = sdf.groupby(['full_name','race'], as_index=False)['race'].agg(['count'])
# creating a pivot table so that each name has a count of the # of races with that last name
gdf = gdf.pivot_table(values='count', columns='race', index='full_name')

# Converting NaN to zeros since that means there is no one that identifies with that race with that last name
gdf = gdf.fillna(0)

gdf['total_n'] = gdf.sum(axis=1)
gdf.reset_index(inplace=True)
gdf.iloc[:, 1:-1] = gdf.iloc[:, 1:-1].div(gdf.total_n, axis=0)

gdf

races = sorted(sdf.race.unique().tolist())
print(races)

def get_race_idx(val, races):
    race_idx = races.index(val)
    return race_idx

# for one set of analysis, we define 'true race/ethincity' = where max prob (so modal race = true race)
gdf['race'] = gdf[races].idxmax(axis=1)
gdf['race_code'] = gdf['race'].apply(lambda c: get_race_idx(c,races))

gdf.head()

gdf.to_csv("train_validation_test/gdf_fullname.csv.gz", index = False, compression="gzip")

## Split dataset in train, validation and test

train_df, rest_df = train_test_split(gdf, test_size=0.2, random_state=42, stratify=gdf['race_code'])
val_df, test_df = train_test_split(rest_df, test_size=0.5, random_state=42, stratify=rest_df['race_code'])

train_df = train_df.reset_index(drop=True)
val_df = val_df.reset_index(drop=True)
test_df = test_df.reset_index(drop=True)

print(train_df.shape)
print(val_df.shape)
print(test_df.shape)

train_df.groupby('race').agg({'full_name':'nunique'})

val_df.groupby('race').agg({'full_name':'nunique'})

test_df.groupby('race').agg({'full_name':'nunique'})

## Save datasets

train_df.to_csv("data/fl_2022_FullName_train.csv.gz", index=False, compression="gzip")
val_df.to_csv("data/fl_2022_FullName_val.csv.gz", index=False, compression="gzip")
test_df.to_csv("data/fl_2022_FullName_test.csv.gz", index=False, compression="gzip")
