==> notebooks/1.0_rf_rapids_fullname_unigrams.ipynb <==
## Installing RAPIDS

!git clone https://github.com/rapidsai/rapidsai-csp-utils.git
!bash rapidsai-csp-utils/colab/rapids-colab.sh stable

!python --version

!python rapidsai-csp-utils/colab/env-check.py

!bash rapidsai-csp-utils/colab/update_gcc.sh
import os
os._exit(00)

## Installing CondaColab.  This will restart your kernel again
import condacolab
condacolab.install()

import condacolab
condacolab.check()

!python rapidsai-csp-utils/colab/install_rapids.py stable

import os
os.environ['NUMBAPRO_NVVM'] = '/usr/local/cuda/nvvm/lib64/libnvvm.so'
os.environ['NUMBAPRO_LIBDEVICE'] = '/usr/local/cuda/nvvm/libdevice/'
os.environ['CONDA_PREFIX'] = '/usr/local'

## Load data using cudf

import cudf as pd


df = pd.read_csv('/content/drive/MyDrive/Colab/ethnicolor/data/fl_reg_name_race_2022.csv.gz')

df.head()

df.describe()

df['race'].unique()

df.dropna(subset=['name_first', 'name_last'], inplace=True)

df.describe()

df.groupby('race').agg({'name_last': 'nunique'})

df['full_name'] = df['name_last'] + ' ' + df['name_first']

df.head()

## Get features using CountVectorizer

from cuml.feature_extraction.text import CountVectorizer

import cuml

cuml.global_settings.output_type

vectorizer = CountVectorizer(analyzer='char')

X = vectorizer.fit_transform(df['full_name'])

X[:10].toarray()

vectorizer.get_feature_names()

len(vectorizer.get_feature_names())

vocab = vectorizer.get_feature_names()

df.head()

X[0].toarray()

import cupy as np

np.unique(df.race.factorize()[0])

df['race_code'] = df.race.factorize()[0]

df.head()

df[['race', 'race_code']].drop_duplicates().sort_values('race_code')

race_id_df = df[['race', 'race_code']].drop_duplicates().sort_values('race_code')

race_id_df

race_to_id = {}
id_to_race = {}

for index, row in race_id_df.to_pandas().iterrows():
    race_to_id[row['race']] = row['race_code']
    id_to_race[row['race_code']] = row['race']

race_to_id

id_to_race

y = df.race_code.to_cupy()

X = X.toarray()

y[0]

y.shape

type(X)

## Preparing for training

from cuml.model_selection import train_test_split

X_train,  X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

from cuml.ensemble import RandomForestClassifier

## Training

#Create a Gaussian Classifier
clf = RandomForestClassifier(split_criterion='entropy', random_state=42, verbose=True)
#Train the model using the training sets y_pred=clf.predict(X_test)
clf.fit(X_train, y_train)

## Accuracy using test dataset

y_pred = clf.predict(X_test)

# Model Accuracy, how often is the classifier correct?
print("Accuracy:", cuml.metrics.accuracy.accuracy_score(y_test, y_pred))

## Save model and vectorizer

import joblib

joblib.dump(clf, "/content/drive/MyDrive/Colab/ethnicolor/fl_voter_name_2022_rf_model_1g.joblib", compress=3)  # compression is ON!
joblib.dump(vectorizer, "/content/drive/MyDrive/Colab/ethnicolor/fl_voter_name_2022_rf_vect_1g.joblib", compress=3)  # compression is ON!
