==> notebooks/5.2_multioutput_regressor_lastname.ipynb <==
### Multi-Output Regressor (DL)

import numpy as np
import pandas as pd

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix
import joblib

from numpy import array, hstack, math
from numpy.random import uniform

import matplotlib.pyplot as plt

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.multioutput import MultiOutputRegressor

%%time
NGRAMS = 2

df = pd.read_csv("data/fl_reg_name_race_2022.csv.gz", engine = "pyarrow")
df.dropna(subset=['name_last'], inplace=True)

# We assume unknown as missing at random
sdf = df[df.race.isin(['unknown']) == False]
del df

# Setting consistent case for names
sdf['name_last'] = sdf.name_last.str.strip().str.lower()

sdf

# Define the mapping for recoding
mapping = {'multi_racial': 'other', 'native_indian': 'other'}

# Recode the 'Category' column using replace()
sdf['race'] = sdf['race'].replace(mapping)
sdf['race'].value_counts()

gdf = sdf.groupby(['name_last','race'], as_index=False)['race'].agg(['count'])
gdf_sm = gdf[(gdf['count'] >= 5) & (gdf['name_last'].str.len() > 1)]
gdf_sm.shape

# creating a pivot table so that each name has a count of the # of races with that last name
gdf_pv = gdf_sm.pivot_table(values='count', columns='race',index='name_last')

# Converting NaN to zeros since that means there is no one that identifies with that race with that last name
gdf_pv = gdf_pv.fillna(0)

# Getting the totals of each last name
gdf_pv['total_n'] = gdf_pv.sum(axis=1)
gdf_pv.reset_index(inplace=True)
gdf_pv

%%time
# Calculate the proportion of people with a particular last name
#  that identify with various races/ethnicity
gdf_pv.iloc[:, 1:] = gdf_pv.iloc[:, 1:].div(gdf_pv.total_n, axis=0)
gdf_pv

gdf_pv.shape

X = gdf_pv.name_last
y = gdf_pv.iloc[:, 1:6]
y

xtrain, xtest, ytrain, ytest=train_test_split(X, y, test_size=0.15)
print("xtrain:", xtrain.shape, "ytrain:", ytrain.shape)
print("xtest:", xtest.shape, "ytest:", ytest.shape)

vect = CountVectorizer(analyzer='char', max_df=0.3, min_df=3, ngram_range=(1, NGRAMS), lowercase=False) 

X_combined_transformed = vect.fit_transform(X)

# Transform the training data
xtrain = X_combined_transformed[:xtrain.shape[0]]

# Transform the test data
xtest = X_combined_transformed[xtrain.shape[0]:]

import torch
import torch.nn as nn
import torch.optim as optim
from sklearn.preprocessing import LabelEncoder

input_size = xtrain.toarray().shape[1]
output_size = ytrain.shape[1]

# Convert the data to PyTorch tensors
input_data = torch.tensor(xtrain.toarray(), dtype=torch.float32)
output_data = torch.tensor(ytrain.values, dtype=torch.float32)

testx_data = torch.tensor(xtest.toarray(), dtype=torch.float32)
testy_data = torch.tensor(ytest.values, dtype=torch.float32)

# Define the neural network model
class MultiOutputRegression(nn.Module):
    def __init__(self, input_size, output_size):
        super(MultiOutputRegression, self).__init__()
        self.linear1 = nn.Linear(input_size, 128)
        self.relu1 = nn.ReLU()
        self.linear2 = nn.Linear(128, 64)
        self.relu2 = nn.ReLU()
        self.linear3 = nn.Linear(64, output_size)
        
    def forward(self, x):
        x = self.linear1(x)
        x = self.relu1(x)
        x = self.linear2(x)
        x = self.relu2(x)
        x = self.linear3(x)
        return x

# Set random seed for reproducibility
torch.manual_seed(42)

# Define the model
model = MultiOutputRegression(input_size, output_size)

# Define the loss function
criterion = nn.MSELoss()

# Define the optimizer
optimizer = optim.Adam(model.parameters(), lr=0.01)

num_epochs = 5000
best_loss = float('inf')
patience = 2000
counter = 0

for epoch in range(num_epochs):
    # Forward pass
    outputs = model(input_data)
    loss = criterion(outputs, output_data)

    # Backward and optimize
    optimizer.zero_grad()
    loss.backward()
    optimizer.step()

    # Print progress
    if (epoch+1) % 100 == 0:
        print(f'Epoch [{epoch+1}/{num_epochs}], Loss: {loss.item():.4f}')

    # Calculate test loss
    with torch.no_grad():
        test_outputs = model(testx_data)
        test_loss = criterion(testy_data, test_outputs)

    # Check for early stopping
    if test_loss < best_loss:
        best_loss = test_loss
        counter = 0
    else:
        counter += 1
        if counter >= patience:
            print(f'Early stopping at epoch {epoch+1}')
            break

ytrain.shape[1]

test_data = xtest.toarray()
test_data = torch.tensor(test_data, dtype=torch.float32)
predictions = model(test_data)

# Convert predictions to numpy array
predictions = predictions.detach().numpy()

(np.argmax(ytest.values, axis = 1) == np.argmax(predictions, axis = 1)).mean()

pd.crosstab(np.argmax(ytest.values, axis = 1), np.argmax(predictions, axis = 1))

normed = predictions/np.sum(predictions, axis = 1).reshape(-1, 1)

(np.argmax(ytest.values, axis = 1) == 3).mean()

print("white normed MSE:%.4f" % mean_squared_error(ytest.iloc[:,3], normed[:,3]))

print("asian MSE:%.4f" % mean_squared_error(ytest.iloc[:,0], predictions[:,0]))
print("hispanic MSE:%.4f" % mean_squared_error(ytest.iloc[:,1], predictions[:,1]))
print("nh_black MSE:%.4f" % mean_squared_error(ytest.iloc[:,2], predictions[:,2]))
print("nh_white MSE:%.4f" % mean_squared_error(ytest.iloc[:,3], predictions[:,3]))

print("base asian MSE:%.4f" % mean_squared_error(ytest.iloc[:,0], [np.mean(ytest.iloc[:,0]) for _ in range(30396)]))
print("base white MSE:%.4f" % mean_squared_error(ytest.iloc[:,3], [np.mean(ytest.iloc[:,3]) for _ in range(30396)]))

torch.save(model, 'v2/fl_last_name_multioutput.pt')
