==> notebooks/6.1_knn_fullname.ipynb <==
## Predict Race/Ethnicity from Unseen Full Name Using KNN LSH MinHash Parallelized

Using the Florida Voting Registration data, we build a knn classifier that predicts the ethnicity of an **unseen** name. We estimate approximate jaccard distance between names using bi-char tokens of the name.

import collections
import multiprocessing as mp
from multiprocessing import Pool
from functools import partial

import tempfile
import pickle
import os

import numpy as np
import pandas as pd

from tqdm import tqdm

from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer                                                             
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix
from sklearn.metrics.pairwise import cosine_similarity
from tqdm import tqdm

from datasketch import MinHashLSHForest, MinHash

# Florida voter
train_df = pd.read_csv('data/fl_2022_FullName_train.csv.gz', usecols=['full_name', 'race'])
val_df = pd.read_csv('data/fl_2022_FullName_val.csv.gz', usecols=['full_name', 'race'])
test_df = pd.read_csv('data/fl_2022_FullName_test.csv.gz', usecols=['full_name', 'race'])

train_df = train_df.groupby('race', group_keys=False).apply(lambda x: x.sample(frac=.7, random_state=10))
train_df.reset_index(inplace=True)
train_df.drop('index', axis=1, inplace=True)
train_df.shape

print('Training set size: {}'.format(train_df.shape))
print('Validation set size: {}'.format(val_df.shape))
print('Test set size: {}'.format(test_df.shape))

## Find the best K

def get_bigrams(text):
    bigrams = []
    for i in range(len(text) - 1):
        bigram = text[i:i + 2]
        bigrams.append(bigram)
    return bigrams

def create_minhash(row):
    set_of_bigrams = get_bigrams(row)
    minhash = MinHash(num_perm=num_perm)
    for term in set_of_bigrams:
        minhash.update(term.encode('utf-8'))
    return minhash

num_processes = mp.cpu_count()
num_perm = 256  # Specify the desired number of permutations

with Pool(processes=num_processes) as pool:
    minhashes = list(tqdm(pool.imap(create_minhash, train_df.full_name), total=train_df.shape[0], desc='Processing rows', unit='row'))

# Index the forest
forest = MinHashLSHForest(num_perm=num_perm)
for i, minhash in enumerate(minhashes):
    key = f"{i}"
    forest.add(key, minhash)
forest.index()

# For prototyping
# ---------------------
# Query the forest to find the k-nearest neighbors for a given query document
query = test_df.full_name[0]
query_minhash = MinHash(num_perm=256)
query_terms = get_bigrams(test_df.full_name[0])
for j in query_terms:
    query_minhash.update(j.encode('utf-8'))

# Find the nearest neighbors using KNN search
k = 5  # Number of nearest neighbors to retrieve
result = forest.query(query_minhash, k)

print(query)
# Step 12: Print the nearest neighbors
for key in result:
    print("Nearest neighbor:", train_df.full_name[int(key)])

def estimate_knn_performance(forest, test_df, k_values, batch_size=100):
    performance = {}

    for k in k_values:
        correct_predictions = 0
        total_examples = len(test_df)
        predicted_labels = []
        true_labels = []
        
        with tempfile.TemporaryDirectory() as temp_dir:
            num_batches = int(np.ceil(len(test_df) / batch_size))
            batches = np.array_split(test_df, num_batches)
            
            for batch_idx, batch in tqdm(enumerate(batches), desc=f'Processing queries (k={k})', total=num_batches):
                result_file = f'{temp_dir}/results_batch_{batch_idx}.pickle'
                process_query_batch(batch, k, result_file)
                
                # Read batch results from disk
                with open(result_file, 'rb') as f:
                    batch_predicted_labels, batch_true_labels = pickle.load(f)
                
                predicted_labels.extend(batch_predicted_labels)
                true_labels.extend(batch_true_labels)
                
                # Delete the result file after reading
                os.remove(result_file)
        
        correct_predictions = sum(pred == true for pred, true in zip(predicted_labels, true_labels))
        accuracy = correct_predictions / total_examples
        performance[k] = accuracy
        
        report = classification_report(pd.Series(true_labels).replace(replacement).to_list(),
                                           pd.Series(predicted_labels).replace(replacement).to_list(),
                                           zero_division='warn')
        print(f"Classification Report (k={k}) - Batch {result_file}:\n{report}\n")
    

    return performance

def process_query_batch(rows, k, result_file):
    batch_size = len(rows)
    query_minhashes = [MinHash(num_perm=256) for _ in range(batch_size)]
    query_terms_list = [get_bigrams(row['full_name']) for _, row in rows.iterrows()]
    
    for i, query_terms in enumerate(query_terms_list):
        for term in query_terms:
            query_minhashes[i].update(term.encode('utf-8'))
    
    result_batch = []
    for query_minhash in query_minhashes:
        result = forest.query(query_minhash, int(k))
        result_batch.append(result)
    
    predicted_labels = []
    true_labels = []
    
    for result, (_, row) in zip(result_batch, rows.iterrows()):
        label_counts = {}

        if result:
            for j in result:
                try:
                    index = int(j)
                    predicted_label = train_df.race[index]

                    if predicted_label in label_counts:
                        label_counts[predicted_label] += 1
                    else:
                        label_counts[predicted_label] = 1

                except (KeyError, ValueError):
                    continue

            # Determine the predicted label based on the majority count
            predicted_label = max(label_counts, key=label_counts.get)
        else:
            # Default to most frequent label from the training data if query result is empty
            predicted_label = most_frequent_label(train_df.race)

        predicted_labels.append(predicted_label)
        true_labels.append(row['race'])
    
    # Write batch results to disk
    with open(result_file, 'wb') as f:
        pickle.dump((predicted_labels, true_labels), f)

replacement = {0: 'asian', 1: 'hispanic', 3: 'nh_black', 4: 'nh_white', 5: 'other'}

%%time
k_values = [10, 25]
performance = estimate_knn_performance(forest, val_df, k_values)
print(performance)

estimate_knn_performance(forest, test_df, [10])
