==> notebooks/9.1_stanford_dime.ipynb <==
# Contrib data from Standford DIME

## Downloading the contribuation list from Stanford DIME
https://data.stanford.edu/dime

import gdown

# downloading the contribDB_2014.csv of size 10GB
# https://drive.google.com/file/d/0B5SFXStyeVhrb1JnaVYzT2p6WWM/view?usp=drive_link&resourcekey=0-WEXQFkUpA46AEUev_ucR7w

url = 'https://drive.google.com/uc?id=0B5SFXStyeVhrb1JnaVYzT2p6WWM'
output = 'contribDB_2014.csv'
gdown.download(url, output, quiet=False)

# peaking first 10 rows
!head -10 contribDB_2014.csv

import pandas as pd

# take a small chunk, to peak data
!head -1000 data/dime/contribDB_2014.csv > data/dime/contribDB_2014_min.csv

df = pd.read_csv('data/dime/contribDB_2014_min.csv', engine='python')

df.head()

df.columns

## Dropping rest of the columns to reduce dataset size

!du -sh data/dime/contribDB_2014.csv

# get only first name and last names
!awk -F, '{print $4","$8","$9}' data/dime/contribDB_2014.csv > data/dime/contribDB_2014_names.csv

!du -sh data/dime/contribDB_2014_names.csv

!wc -l data/dime/contribDB_2014_names.csv

!head -10 data/dime/contribDB_2014_names.csv

!sed -i 's/"//g' data/dime/contribDB_2014_names.csv

!head -10 data/dime/contribDB_2014_names.csv

df = pd.read_csv('data/dime/contribDB_2014_names.csv', engine='python')

df.shape

## Dropping Nan

df = df.dropna()

df.shape

df.head()

## Dropping orgs

df = df.drop(df[df['contributor.lname'].str.contains("&|employee|associates|\d",regex=True)].index)

## Make title case

df['contributor.lname'] = df['contributor.lname'].str.title()
df['contributor.fname'] = df['contributor.fname'].str.title()

df.head()

df.shape

## Remove special characters

df['contributor.lname'] = df['contributor.lname'].str.replace("[^a-zA-Z' -]", '', regex=True)
df['contributor.fname'] = df['contributor.fname'].str.replace("[^a-zA-Z' -]", '', regex=True)

df['contributor.lname'] = df['contributor.lname'].str.strip()
df['contributor.fname'] = df['contributor.fname'].str.strip()

## Droping names less than 2

df = df.drop(df[df['contributor.lname'].str.len() < 2].index)
df = df.drop(df[df['contributor.fname'].str.len() < 2].index)

df.shape

df[df['contributor.lname'] == 'Smith']

df[df['contributor.fname'] == 'Smith']

## Majority of lnames and fnames are swapped

df['name_first'] = df['contributor.lname']
df['name_last'] = df['contributor.fname']

## Removing middle names

df['name_first'] = df.name_first.apply(lambda x: x.split()[0])

## find first names that has lastname ans swap

df.head()

## Drop if first and last names are matched.

df[df['name_first'] == df['name_last']]

df = df.drop(df[df['name_first'] == df['name_last']].index)

df = df.drop(df[df['name_first'].str.len() < 2].index)
df = df.drop(df[df['name_last'].str.len() < 2].index)

df.head()

last_names = df['name_last'].value_counts().loc[lambda x: x>10000].reset_index()['name_last'].values
first_names = df['name_first'].value_counts().loc[lambda x: x>10000].reset_index()['name_first'].values

df['full_name'] = df['name_last'] + ' ' + df['name_first']

df.head()

df['amount'] = df['amount'].apply(lambda x: pd.to_numeric(x, errors='coerce'))

df.shape

df.dropna(inplace=True)

df.shape

df.groupby(['full_name','amount'], as_index=False)['amount'].agg(['sum'])

sdf = df.groupby(['full_name'], as_index=False)['amount'].agg(['sum'])

sdf.head()

cdf = df.groupby(['full_name'], as_index=False)['amount'].agg(['count'])

cdf.head()

print(cdf.shape)
print(sdf.shape)

x = "A Anderson Nate"

final_df = pd.merge(sdf, cdf, on=['full_name'])

final_df.shape

final_df['full_name'] = final_df['full_name'].apply(lambda x : ' '.join([i for i in x.split() if len(i) > 2]))

final_df.head()

final_df.shape

final_df.dropna(inplace=True)

final_df.shape

final_df = final_df.drop(final_df[final_df['full_name'].str.contains("Professional|Advocate|Employee|Assoc|Sponsored",regex=True)].index)

final_df.shape

final_df.to_csv("data/dime/contribDB_2014_final.csv", index=False)
