==> notebooks/11.1_rf_fullname_interp.ipynb <==
### RF ML Model (to aid basic interpretation)

import numpy as np
import pandas as pd
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix
import joblib

# Florida voter
df = pd.read_csv("train_validation_test/fl_2022_fullname.csv.gz",
                 usecols=['full_name', 'race', 'race_code'])

# stratified sample
proto_df = df.groupby('race', group_keys=False).apply(lambda x: x.sample(frac=.2, random_state=10))
proto_df.reset_index(inplace=True)
proto_df.drop('index', axis=1, inplace=True)
proto_df.shape

race_id_df = proto_df[['race', 'race_code']].drop_duplicates().sort_values('race_code')
race_to_id = dict(race_id_df.values)
id_to_race = dict(race_id_df[['race_code', 'race']].values)
id_to_race

X = proto_df.full_name
y = proto_df.race_code

NGRAMS = 2
#vect = TfidfVectorizer(analyzer='char', sublinear_tf=True, min_df=5, norm='l2', ngram_range=(1, NGRAMS), lowercase=False)
vect = CountVectorizer(analyzer='char', max_df=0.3, min_df=5, ngram_range=(1, NGRAMS), lowercase=False) 

features = vect.fit_transform(X).toarray()
labels = y
features.shape

from sklearn.feature_selection import chi2

N = 5
for race_code, race in id_to_race.items():
  features_chi2 = chi2(features, y == race_code)
  indices = np.argsort(features_chi2[0])
  feature_names = np.array(vect.get_feature_names_out())[indices]
  bigrams = [v for v in feature_names if len(v) == 2]
  print("# '{}':".format(race))
  print("  . Most correlated bigrams:\n       . {}".format('\n       . '.join(bigrams[-N:])))

# Split train and test dataset
X_train,  X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=21, stratify=y)

X_train_vect = vect.fit_transform(X_train)
feature_names = vect.get_feature_names_out()

%%time

from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics

clf = RandomForestClassifier(n_estimators = 4,
                             max_samples  = .5, 
                             criterion    = 'entropy', 
                             random_state = 21)
clf.fit(X_train_vect, y_train)

X_test = vect.transform(X_test).toarray()
y_pred = clf.predict(X_test)

# Model Accuracy, how often is the classifier correct?
print("Accuracy:", metrics.accuracy_score(y_test, y_pred))

from sklearn.inspection import permutation_importance

# Asian
class_index = 0  
result = permutation_importance(clf, vect.transform(X_train).toarray(), y_train == class_index, n_repeats=1, max_samples = 1000, random_state=42)

# Get the feature importance scores
importance_scores = result.importances_mean

# Get the indices of features sorted by importance in descending order
sorted_indices = importance_scores.argsort()[::-1]

# Print the feature importance scores and corresponding feature names
for idx in sorted_indices[0:20]:
    print(f"Feature: {feature_names[idx]}, Importance Score: {importance_scores[idx]}")

# Hispanic
class_index = 1  
result = permutation_importance(clf, vect.transform(X_train).toarray(), y_train == class_index, n_repeats=1, max_samples = 1000, random_state=42)

# Get the feature importance scores
importance_scores = result.importances_mean

# Get the indices of features sorted by importance in descending order
sorted_indices = importance_scores.argsort()[::-1]

# Print the feature importance scores and corresponding feature names
for idx in sorted_indices[0:20]:
    print(f"Feature: {feature_names[idx]}, Importance Score: {importance_scores[idx]}")

## Confusion Matrix

target_names = list(df.race.astype('category').cat.categories)
print(classification_report(y_test, y_pred, target_names=target_names))
print(confusion_matrix(y_test, y_pred))

## Save model

joblib.dump(clf, "models/rf_fullname_interp.joblib", compress=3)  # compression is ON!
