==> notebooks/12.0_census_data_prep.ipynb <==
import pandas as pd

df = pd.read_csv("census_2010_processed.csv")

df.head()

df = df[['name','ethni_act']]

df['race_code'] = df.ethni_act.factorize()[0]

df.head()

df = df.rename(columns={'ethni_act': 'race','name':'name_last'})

df.head()

from sklearn.model_selection import train_test_split

train_df, rest_df = train_test_split(df, test_size=0.2, random_state=42, stratify=df['race_code'])
val_df, test_df = train_test_split(rest_df, test_size=0.5, random_state=42, stratify=rest_df['race_code'])

print(train_df.shape)
print(val_df.shape)
print(test_df.shape)

train_df.to_csv("data/census_ln_train.csv",index=False)
val_df.to_csv("data/census_ln_val.csv",index=False)
test_df.to_csv("data/census_ln_test.csv",index=False)

!find . -name "census_ln_*.csv" | xargs du -sh

## Augmenting data

df.head()

ln_train_df = pd.read_csv("data/fl_2022_LastName_train.csv.gz")
ln_val_df = pd.read_csv("data/fl_2022_LastName_val.csv.gz")
ln_test_df = pd.read_csv("data/fl_2022_LastName_test.csv.gz")

ln_df = pd.concat([ln_train_df,ln_val_df,ln_test_df])

ln_df = ln_df[['name_last', 'race', 'race_code']]

ln_df.shape

df.shape

df_all = ln_df.merge(df, on=['name_last','name_last'], how='left', indicator=True)

df_all.head()

not_in_df = df_all[df_all['_merge'] == 'left_only'][['name_last','race_x','race_code_x']]

not_in_df = not_in_df.rename(columns={'race_x':'race', 'race_code_x':'race_code'})

not_in_df.head()

## Sync the race_code
race_id_df = df[['race', 'race_code']].drop_duplicates().sort_values('race_code')
race_to_id = dict(race_id_df.values)

race_to_id

not_in_df['race_code'] = not_in_df.apply(lambda x: race_to_id[x['race']] ,axis=1)

not_in_df.head()

final_df = pd.concat([train_df, not_in_df])

final_df.shape

final_df.to_csv("data/census_ln_train_w_ln_ds.csv",index=False)
